+++ setup.py
@@ -205,7 +205,13 @@ setup(name='montblanc',
       packages=[
       	'montblanc',
       	'montblanc.examples',
-        'montblanc.ext'],
+        'montblanc.ext',
+        'montblanc.impl',
+        'montblanc.impl.biro',
+        'montblanc.impl.biro.v1',
+        'montblanc.impl.biro.v1.gpu',
+        'montblanc.impl.biro.v1.cpu',
+        'montblanc.impl.biro.v1.test',],
       install_requires=[
       	'numpy',
       	'pycuda',
